pb 상품 페이지만 앱바 다홍색으로 표시
