--- store/lib/src/adapter/presentation/modules/pb_product/bindings/pb_product_bindings.dart
@@ -1,10 +1,0 @@
-import 'package:get/get.dart';
-
-import '../controller/pb_product_controller.dart';
-
-class PbProductBindings extends Bindings {
-  @override
-  void dependencies() {
-    Get.put(PbProductController());
-  }
-}
